+++ src/global/redux/index.js
@@ -1,5 +1,5 @@
 import auth from 'global/redux/auth/slice';
 import cart from 'global/redux/cart/slice';
-import address from 'global/redux/user/slice';
+import address from 'global/redux/address/slice';
 
 export { address, auth, cart };
